+++ src/main/resources/application.yaml
@@ -9,10 +9,10 @@ management:
     web:
       base-path: /
       exposure:
-        include: health, info, prometheus
+        include: health, info
 
 springdoc:
-  packagesToScan: uk.gov.hmcts.reform.demo.controllers
+  packagesToScan: uk.gov.hmcts.reform.cosapi.controllers
 
 #If you use a database then uncomment below lines and update db properties accordingly
 spring:
@@ -45,3 +45,9 @@ spring:
 azure:
   application-insights:
     instrumentation-key: ${rpe.AppInsightsInstrumentationKey:00000000-0000-0000-0000-000000000000}
+
+idam:
+  s2s-auth:
+    url: ${SERVICE_AUTH_PROVIDER_URL:http://localhost:4552}
+    secret: ${S2S_SECRET:AAAAAAAAAAAAAAAA}
+    microservice: fis_cos_api
